Move 'FetchKeys()' function into a subpkg (#435)
* Move 'FetchKeys()' function into a subpkg (#433)

Fixes #433


* refactor 'token/keycloak/keys' pkg to 'token/jwt'


* Rename jwt to jwk

* Rename jwt to jwk
