+++ token/jwk/doc.go
@@ -1,0 +1,2 @@
+// Package token is used to manage JSON Web Keys
+package jwk
